back-app/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
GeoAlchemy2==0.8.4
geopandas==0.8.2
pandas==1.2.2
SQLAlchemy==1.3.23
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
wincertstore==0.2
